Decouple and encapsulate
