+++ COVID19Py/__init__.py
@@ -1,1 +1,3 @@
 from .covid19 import COVID19
+from .overview import OverviewBased
+from .location import LocationBased
