+++ COVID19Py/overview.py
@@ -1,0 +1,27 @@
+class OverviewBased:
+    def getAll(self, timelines=False):
+        self._update(timelines)
+        return self.latestData
+
+    def getLatestChanges(self):
+        changes = None
+        if self.previousData:
+            changes = {
+                "confirmed": self.latestData["latest"]["confirmed"] - self.previousData["latest"]["confirmed"],
+                "deaths": self.latestData["latest"]["deaths"] - self.previousData["latest"]["deaths"],
+                "recovered": self.latestData["latest"]["recovered"] - self.previousData["latest"]["recovered"],
+            }
+        else:
+            changes = {
+                "confirmed": 0,
+                "deaths": 0,
+                "recovered": 0,
+            }
+        return changes
+
+    def getLatest(self) -> List[Dict[str, int]]:
+        """
+        :return: The latest amount of total confirmed cases, deaths, and recoveries.
+        """
+        data = self._request("/v2/latest")
+        return data["latest"]
